+++ src/components/3d/index.js
@@ -1,0 +1,3 @@
+export { default as CouncilScene } from './CouncilScene';
+export { default as AgentModel } from './AgentModel';
+export { default as CouncilTable } from './CouncilTable';